Flatten simple commands
